+++ objects/src/block/header.rs
@@ -44,9 +44,6 @@ pub struct BlockHeader {
 }
 
 impl BlockHeader {
-    /// The length of an epoch expressed as a power of two. `2^(EPOCH_LENGTH_EXPONENT)` is the
-    /// number of blocks in an epoch.
-
     /// Creates a new block header.
     #[allow(clippy::too_many_arguments)]
     pub fn new(
@@ -132,7 +129,7 @@ impl BlockHeader {
 
     /// Returns the epoch to which this block belongs.
     ///
-    /// This is the block number shifted right by [`Self::EPOCH_LENGTH_EXPONENT`].
+    /// This is the block number shifted right by [`BlockNumber::EPOCH_LENGTH_EXPONENT`].
     pub fn block_epoch(&self) -> u16 {
         self.block_num.block_epoch()
     }
@@ -274,12 +271,10 @@ impl Deserializable for BlockHeader {
 
 /// BLOCK NUMBER
 
-/// Holds `u32` type to signify Block Number
-
-#[derive(Debug, Eq, PartialEq, Copy, Clone, PartialOrd, Ord, Hash)]
 /// A convenience wrapper around a `u32` representing the number of a block.
 ///
 /// Each block has a unique number and block numbers increase monotonically by `1`.
+#[derive(Debug, Eq, PartialEq, Copy, Clone, PartialOrd, Ord, Hash)]
 pub struct BlockNumber(u32);
 
 impl Serializable for BlockNumber {
@@ -305,18 +300,24 @@ impl fmt::Display for BlockNumber {
 }
 
 impl BlockNumber {
+    /// The length of an epoch expressed as a power of two. `2^(EPOCH_LENGTH_EXPONENT)` is the
+    /// number of blocks in an epoch.
+    ///
     /// The epoch of a block can be obtained by shifting the block number to the right by this
     /// exponent.
     pub const EPOCH_LENGTH_EXPONENT: u8 = 16;
 
+    /// Creates the [`BlockNumber`] corresponding to the epoch block for the provided `epoch`.
     pub const fn from_epoch(epoch: u16) -> BlockNumber {
         BlockNumber((epoch as u32) << BlockNumber::EPOCH_LENGTH_EXPONENT)
     }
 
+    /// Returns the epoch to which this block number belongs.
     pub const fn block_epoch(&self) -> u16 {
         (self.0 >> BlockNumber::EPOCH_LENGTH_EXPONENT) as u16
     }
 
+    /// Returns the block number as a `u32`.
     pub fn as_u32(&self) -> u32 {
         self.0
     }